pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurokg-project"
version = "0.1.0"
description = "Split-table NeuroKG scaffolding"
requires-python = ">=3.10"
dependencies = [
  "psycopg2-binary>=2.9",
  "pyyaml>=6.0",
  "pandas>=2.0",
  "flask>=3.0",
  "PySide6>=6.8"
]
